Install gem panolint only on Ruby 2.7
We only run linters in CI on Ruby 2.7, and the dependencies
of the panolint gem have broken (hopefully temporarily) on
JRuby now.

We'd like to match our CI build dependencies up with the tests
that we run, so that we don't get test failures from dependency
incompatibilities with Ruby versions that we won't be running them
on.