Allow keyword arguments in initializers when including MemoWise
Currently on ruby 3.0, prepending MemoWise into a class which accepts keyword arguments
in its initializer causes an ArgumentError to be thrown:

ArgumentError (wrong number of arguments (given 1, expected 0; required keyword: :kwarg))

This change allows both positional and keyword arguments to initializers in classes which
use MemoWise.  This is compatible with versions before 3.0 as well.