Fix Marshal serialization
As pointed out by @honigc in #137, the use of a default proc in the
core MemoWise Hash instance makes serialization using Marshal impossible.

This PR replaces the implementation with a simple Hash, and sets the
default values along the API code paths that previously relied on the
default value.

We cover the case in #reset_memo_wise where specific args are passed,
to avoid regression of a Nil error when attempting to reset args that
are not yet memoized, as this error would become possible now that
there is no default value proc in the Hash.

Closes #137

Also:
  * Update CHANGELOG.md - added support for Marshal serialization
  * Update benchmarks in README.md - no significant changes