Remove broken Github Action workflows
This PR removes the following Github Actions workflows:
  * auto-approve-dependabot.yml
  * remove-needs-qa.yml

We intend to replace the auto-approve approach to dependabots
with a new solution that works with the new github-native
dependabot in as secure a fashion as we can. The existing
solution doesn't work for a few reasons.

I don't believe we need the "remove Needs-QA" workflow for
this repository, as it is not a webapp that is auto-deployed
using the workflows that look for this label.