Decorate .instance_method to proxy original params
In Panorama spec helper code 'initialized_double', we observed
that MemoWise interrupts the use of Ruby reflection on method
parameters, by replacing with delegating methods that have overly
generic signatures, like:

    def initialize(...)

    def slow_method(*arg, **kwargs, &block)

After investigation, we had determined that it was not worth
the complexity cost, and potential fragility, to pursue a solution
where we fully re-create the original method parameter signatures
in our delegating implementations.

However, we had decided to explore a more minimalistic alternative.

This PR is a minimalistic alternative, which narrowly overrides only
the implementation of {Module#instance_method} so that, in the case
of either #initialize or a method that has had .memo_wise called on it,
we look up the original method and proxy the original result value of
calling UnboundMethod#parameters on it.