Improve and document benchmarking for truthy optimizations
Before this commit, our benchmarks would lead someone to
change our cache-lookup implementation back to using `fetch`,
because the performance penalty for falsey method results
overshadowed the performance penalty of the `fetch` method
call.

We believe optimizing for truthy results is a better approach,
so this commit updates our benchmarks to reflect that decision
(and document it), making `fetch` less attractive in our
benchmarks.

NOTE: These benchmark results are specific to if we had
turned off our `.hash` optimization in both cases, because
that optimization is likely to be removed and reduces the
overhead of the conditional.