Fix thread-safety for concurrent uncached calls
This updates the sentinel after updating the cache, and check the sentinel before checking the cache.  Otherwise, if there are two concurrent calls for an uncached value in separate threads, a thread-safety issue is possible:

1. Thread 1 sets sentinel value to true
2. Thread 2 gets cached value (nil is not set)
3. Thread 2 checks sentinel value (is true)
4. Thread 2 returns nil instead of correct value
5. Thread 1 sets cached value

This changes the behavior so that in the case of concurrent calls for an uncached value, the original method is called twice instead of in invalid value being returned.

This change may negatively affect performance for cached calls, I haven't benchmarked or tested this.