Document concurrent memoization of zero-arg methods
Document, via both README and specs, that when two threads concurrently
call a zero-arg method which is in the unmemoized state, they can return
two different results.

For example, this could occur if we are memoizing:
  * database query result
  * filesystem read
  * current time
  * a random number

This PR introduces a test which probabilistically reproduces the
issue, as well as documentation that this is the expected behavior.

Disables a test on TruffleRuby in CI, as it takes multiple seconds
to observe the different return values there.

Closes #184