Revert internal data structure to simple hash
Previous versions of `MemoWise` used an array as the
internal cache data structure, where each memoized method
uses a unique array index. However, complex inheritance
structures make it easy to have multiple methods accidentally
share the same array index on the same object, resulting
in nondeterministic and incorrect behavior.

While we may explore that optimization again in the future,
for the time being we are switching to a hash data structure
keyed on method name, which is safer because two methods
cannot share the same name in Ruby.