Fix race conditions in methods with arguments
This commit fixes race conditions found in memoized methods
that take arguments. This comes at a slight performance cost
for truthy memoized values (due to switching from `Hash#[]`
to the slightly slower `Hash#fetch`) but will be faster when
memoized values are falsey.

This change removes the "truthy optimization" leading to these
race conditions and as a result allows us to simplify our
benchmark code as we no longer need to benchmark truthy and
falsey values separately.