Use nested hashes for multiple required arguments
This commit optimizes methods with multiple required arguments to use
nested hashes to store the memoized value (so each argument accesses
the next layer in the hash) rather than a single-layer hash with an
array key.

This is more performant because multiple `#hash` calls on simpler
data structures are faster than one `#hash` lookup on all of those
elements within an array, and in addition we avoid the cost of
initializing an array on every lookup.

Closes #244