Fixed a bug that overwrote existing self.included method definitions.
There is a hack that dynamically defines self.inherited when memo_wise is called, but there was a bug that did not consider the case where an existing self.inherited existed, so this has been fixed.
And I also fixed a bug that defined an `#inherited` method on the instance.