Add CI/benchmarks for Ruby 3.3
This commit updates our CI to run tests and
benchmarks on Ruby 3.3. Note that it does not
yet run CI benchmarks using YJIT.

As a part of this change, the `dokaz` gem
is pinned to run on Ruby 3.2 instead of the
latest Ruby version because the gem does
not work in 3.3+. If that changes, we can
go back to running it on the latest supported
Ruby. See: https://github.com/zverok/dokaz/issues/3