Drop support for Ruby 2.7 and 3.0, add 3.4 test coverage
This commit removes support for Ruby versions below 3.1 (which are EOL),
so our tests can pass with the improvement made in
4c6e01e29feb33808aa8c4f235e092195d41c794. In addition, since TruffleRuby
does not yet support the behavior change from Ruby 3.1 which allows that
improvement to work, certain tests are skipped in TruffleRuby until they
implement that change.

While making these changes, this commit adds Ruby 3.4 to our test and
benchmark suite.